refactor: re-organize the project
