+++ bins/reth-crawler/src/crawler/listener/update_listener.rs
@@ -1,17 +1,15 @@
+use crate::p2p::handshake_p2p;
 use chrono::Utc;
 use futures::StreamExt;
-use reth_crawler_db::PeerDB;
-use reth_crawler_p2p::{handshake_p2p, save_peer};
-use reth_crawler_types::PeerData;
+use ipgeolocate::{Locator, Service};
+use once_cell::sync::Lazy;
+use reth_crawler_db::{save_peer, PeerDB, PeerData};
 use reth_discv4::{DiscoveryUpdate, Discv4};
 use reth_dns_discovery::{DnsDiscoveryHandle, DnsNodeRecordUpdate};
+use reth_primitives::{mainnet_nodes, NodeRecord};
 use secp256k1::SecretKey;
 use std::time::Instant;
 use tokio::sync::mpsc::UnboundedSender;
-
-use ipgeolocate::{Locator, Service};
-use once_cell::sync::Lazy;
-use reth_primitives::{mainnet_nodes, NodeRecord};
 
 pub static MAINNET_BOOT_NODES: Lazy<Vec<NodeRecord>> = Lazy::new(mainnet_nodes);
