+++ bins/reth-crawler/src/main.rs
@@ -1,4 +1,5 @@
 mod crawler;
+mod p2p;
 use clap::{Args, Parser, Subcommand};
 use crawler::CrawlerFactory;
