--- crates/db/src/lib.rs
@@ -1,4 +1,0 @@
-pub mod db;
-
-// Re-exports
-pub use db::PeerDB;
