--- crates/p2p/src/lib.rs
@@ -1,7 +1,0 @@
-pub mod p2p;
-
-// Re-exports
-pub use p2p::append_to_file;
-pub use p2p::handshake_eth;
-pub use p2p::handshake_p2p;
-pub use p2p::save_peer;
