--- crates/types/Cargo.toml
@@ -1,15 +1,0 @@
-[package]
-name = "reth-crawler-types"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-# reth
-reth-primitives.workspace = true
-
-# serialization
-serde_json.workspace = true
-serde.workspace = true
-
-# aws
-aws-sdk-dynamodb.workspace = true
