--- crates/types/src/lib.rs
@@ -1,7 +1,0 @@
-pub mod types;
-
-// Re-exports
-pub use types::as_string;
-pub use types::as_string_vec;
-pub use types::as_u16;
-pub use types::PeerData;
